Fixed bug in _invite_users_modal_results.html.erb
the bug: when inviting users, ater a new user is generated (if needed) and email is sent, an error occurs in the partial responsible for generating the "all is done" message, so that instead of "all is done" one sees "error inviting users", even though the users have been invited successfully. 
The reason for the error is the mismatch between expected structure of @invite_results and the actual @invite_results.