Fix Protocols.io modal not loading any protocols [SCI-9167]
The JavaScript responsible for loading the list was attached to a
Partial that by itself was loaded using and AJAX request. Therefore,
The JavaScript was not being executed (no page reload). It was thus moved
to the parent partial (loaded on page reload) and called when the required
HTML is appended to the DOM.